+++ src/adhanpy/__init__.py
@@ -1,5 +1,6 @@
 from adhanpy.PrayerTimes import PrayerTimes
 from adhanpy.Qibla import Qibla
+from adhanpy.SunnahTimes import SunnahTimes
 from adhanpy.calculation.CalculationMethod import CalculationMethod
 from adhanpy.calculation.CalculationParameters import CalculationParameters
 from adhanpy.calculation.HighLatitudeRule import HighLatitudeRule
@@ -11,6 +12,7 @@ from adhanpy.data.Prayer import Prayer
 __all__ = [
     "PrayerTimes",
     "Qibla",
+    "SunnahTimes",
     "CalculationMethod",
     "CalculationParameters",
     "HighLatitudeRule",